/sakura: Adding M2M kernel.

+++ sakura/utils.h
@@ -70,9 +70,9 @@ void form_interaction_lists(int **node_codes, int **children_first,
 			    int (*restrict nodes_per_level), 
 			    int (*restrict nodes_per_level2), 
 			    int height);
-void upward_pass(float *X2, int (**Multipole)[MTERM], int **node_codes2, int** c_count2,
-		 int** node_pointers2, int *leaf_populations,
+void upward_pass(float *X2, float (**Multipole)[MTERM], int **node_codes2,
+		 int** c_count2, int** node_pointers2, int *leaf_populations,
 		 float *Xmin, float *Xmax, int node_id, int level);
-void downward_pass(float *X, int (**Local)[LTERM], int **node_codes, int** c_count,
-		   int** node_pointers, int *leaf_populations,
+void downward_pass(float *X, float (**Local)[LTERM], int **node_codes,
+		   int** c_count, int** node_pointers, int *leaf_populations,
 		   float *Xmin, float *Xmax, int node_id, int level);
